[DT-1164] Move some ViewModels into non-compose payments modules
To consolidate payments logic but still keep compose deps separate.
